chore: refactor order structure
