+++ src/tsx/components/cards/IntroductionCard.tsx
