+++ src/tsx/components/cards/StakeholderCard.tsx
